feat: Update dependencies and configurations for InterviewChallenge.ApiService
This commit updates the dependencies and configurations for the InterviewChallenge.ApiService project. The Microsoft.EntityFrameworkCore.SqlServer package is updated to version 8.0.5, and the Microsoft.EntityFrameworkCore.Tools package is updated to version 8.0.5. Additionally, the InterviewChallenge.ApiService.csproj file is modified to include the UserSecretsId and the PackageReference for the updated packages.