feat: Update devcontainer.json to use latest versions of GitHub CLI and PowerShell
This commit updates the devcontainer.json file to use the latest versions of the GitHub CLI and PowerShell packages. The version of GitHub CLI is updated to "latest" and the version of PowerShell is updated to "7.4.2". This ensures that the development container has the most up-to-date tools for GitHub and PowerShell.